humaniod-robotics/humaniod-robotics/api/requirements.txt
fastapi>=0.115.0
openai>=1.54.0
qdrant-client>=1.12.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
uvicorn>=0.32.0
